requirements.txt
requests
python-dotenv
subliminal
beautifulsoup4
polars
duckdb
openai
numpy
pandas
sentence-transformers
textblob